Fixed mistakes: Charges now agree with RuNNer.
Charges for AlAuMgO and carbon-chain agree with RuNNer (so both periodic
and non periodic case should be ok). Changed the neighborlist
construction to include rcutScreen and rcutReal. Furthermore it can be
sorted now, but the "<" operator of the Atom class has now changed,
therefore something in nnp-sfclust is now broken.