Added stage 1 training but prob. still incorrect.
In order to train hardness J (actually training h = sqrt(J)) I included
it to the weights vector where for each element the sequence
a,a,...,a,b,...,b,h is stored. Therefore i changed numWeights and other
corresponding values in case of nnpType 4G. I hope this doesn't break
anything. The code seems to run but the result is not yet verified and
probably incorrect.